% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{DTOutput}
\alias{DTOutput}
\title{Embed a DT DataTable output in a irid tag tree}
\usage{
DTOutput(fn, ...)
}
\arguments{
\item{fn}{A function that produces a DataTable.}

\item{...}{Additional arguments passed to \code{DT::DTOutput()}.}
}
\value{
A irid output node.
}
\description{
Shorthand for \code{Output(DT::renderDT, DT::DTOutput, ...)}. Requires the
\strong{DT} package.
}
