% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irid-package.R
\docType{package}
\name{irid-package}
\alias{irid}
\alias{irid-package}
\title{irid: Component-Based Shiny UI with Fine-Grained Reactivity}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A component-based reactive UI framework for Shiny that binds reactive values directly to DOM attributes, enabling fine-grained updates without re-rendering or 'update*Input' callbacks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://irid.kylehusmann.com}
  \item \url{https://github.com/khusmann/irid}
  \item Report bugs at \url{https://github.com/khusmann/irid/issues}
}

}
\author{
\strong{Maintainer}: Kyle Husmann \email{khusmann@gmail.com}

}
\keyword{internal}
