% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tags.R
\name{irid_dependency}
\alias{irid_dependency}
\title{irid JavaScript dependency}
\usage{
irid_dependency()
}
\value{
An \code{html_dependency} object.
}
\description{
Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} for the client-side irid
runtime (\code{irid.js}).
}
\keyword{internal}
