% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tags.R
\name{irid_id_counter}
\alias{irid_id_counter}
\title{Create a local ID counter for use within a single \code{process_tags} call}
\usage{
irid_id_counter(prefix = "irid")
}
\value{
A function that returns the next ID each time it is called.
}
\description{
Create a local ID counter for use within a single \code{process_tags} call
}
\keyword{internal}
