% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{PlotOutput}
\alias{PlotOutput}
\title{Embed a plot output in a irid tag tree}
\usage{
PlotOutput(fn, ...)
}
\arguments{
\item{fn}{A function that produces a plot.}

\item{...}{Additional arguments passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}
}
\value{
A irid output node.
}
\description{
Shorthand for \code{Output(renderPlot, plotOutput, ...)}.
}
