% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{TableOutput}
\alias{TableOutput}
\title{Embed a table output in a irid tag tree}
\usage{
TableOutput(fn, ...)
}
\arguments{
\item{fn}{A function that produces a table.}

\item{...}{Additional arguments passed to \code{\link[shiny:renderTable]{shiny::tableOutput()}}.}
}
\value{
A irid output node.
}
\description{
Shorthand for \code{Output(renderTable, tableOutput, ...)}.
}
